Milestone 4/5
In Milestone 4 we changed the days into months and days, the time in minutes to minutes and hours, the pictures to be changed when they are clicked on, and the thumbs and updown button to be response so the values can change if they are clicked

Milestone 5: Man made error, I did not realize that milestone 5 was add the values that were created in milestone 4 to index html. For milestone 5 I added the function names to my html for example (Time text, movies, and poster number).